requirements.txt
pymupdf==1.27.1
numpy
pillow